# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SAGE Platform API
  description: |
    RESTful API for the SAGE Platform - a project management and client collaboration system.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - Authentication endpoints: 5 requests per minute
    - General API: 100 requests per minute
    - File uploads: 10 requests per minute

    ## Error Responses
    All errors follow a standard format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: SAGE Support
    email: support@sage-platform.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: https://api.sage-platform.com/api
    description: Production
...
paths:
...
  /leads:
    get:
      tags: [Leads]
      summary: List all leads (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/LeadStatus'
        - name: tier
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 4
      responses:
        '200':
          description: List of leads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLeads'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Leads]
      summary: Create a new lead from intake form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntakeFormData'
      responses:
        '201':
          description: Lead created with tier recommendation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadWithRecommendation'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /projects:
    get:
      tags: [Projects]
      summary: List user's projects
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjects'
...
  /admin/dashboard:
    get:
      tags: [Admin]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    LeadStatus:
      type: string
      enum: [new, contacted, qualified, converted, closed]
    IntakeFormData:
      type: object
      required: [email, name, address, budget, timeline, projectType]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        budget:
          type: string
          enum: [under_500, 500_1500, 1500_3500, 3500_plus, custom]
        timeline:
          type: string
          enum: [asap, 1_2_weeks, 2_4_weeks, 1_2_months, flexible]
        projectType:
          type: string
          enum: [residential, commercial, landscape, interior, mixed]
        hasSurvey:
          type: boolean
        hasDrawings:
          type: boolean
        description:
          type: string
    LeadWithRecommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/LeadStatus'
        recommendation:
          type: object
          properties:
            tier:
              type: integer
            reason:
              type: string
            confidence:
              type: number
            needsManualReview:
              type: boolean
            factors:
              type: array
              items:
                type: object
                properties:
                  factor:
                    type: string
                  impact:
                    type: string
                  weight:
                    type: number
    ProjectStatus:
      type: string
      enum: [pending, active, in_progress, review, completed, on_hold, cancelled]
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        progress:
          type: number
        tierId:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    DashboardStats:
      type: object
      properties:
        leads:
          type: object
          properties:
            total:
              type: integer
            byStatus:
              type: object
        projects:
          type: object
          properties:
            total:
              type: integer
            byTier:
              type: object
            byStatus:
              type: object
        revenue:
          type: object
          properties:
            total:
              type: number
            thisMonth:
              type: number
            lastMonth:
              type: number
    PaginatedLeads:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/LeadWithRecommendation'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PaginatedProjects:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer